fix(geomappanel.tsx): fix error when using the spatial filter tool
Due to reusing the datahover functionality in the spatial filter tool there is thrown an error when
using the tool. In particular the pointerMoveListener function the error is thrown due to null value
in the thisLayer variable. This is fixed by providing an empty object if the thisLayer is null.