fix(markerslayer.tsx): fix refId handling in markers layer
The markers layer compares the data series refId with the refId configured in the options object.
But there was a break in the loop so the right refId is never found. Additionally whenever there is
a meta object in the series the conditional expression is true, regardless if the options meets the
query configured in the options.

fix #32